+++ test-e2e/mocks/mocks-server.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  options: {
+    port: 3100,
+    behavior: "foo",
+    cli: false,
+  },
+};