+++ test-e2e/nodejs/es6/test/index.spec.js
@@ -6,17 +6,33 @@ const {
   readFixture,
   readSettings,
   updateSettings,
+  readAlerts,
+  readAlert,
 } = require("../index");
 
 describe("react-admin-client methods used through node", () => {
-  describe("whe reading about", () => {
+  describe("when reading about", () => {
     it("should return current version", async () => {
       const about = await readAbout();
       expect(about.version).toBeDefined();
     });
   });
 
-  describe("whe reading fixtures", () => {
+  describe("when reading alerts", () => {
+    it("should return one alert", async () => {
+      const alerts = await readAlerts();
+      expect(alerts.length).toEqual(1);
+    });
+
+    it("alert model should exist", async () => {
+      const alerts = await readAlerts();
+      const alertId = alerts[0].id;
+      const alert = await readAlert(alertId);
+      expect(alert.id).toEqual(alertId);
+    });
+  });
+
+  describe("when reading fixtures", () => {
     it("should return current fixtures collection", async () => {
       const fixtures = await readFixtures();
       expect(fixtures.length).toEqual(2);
@@ -37,7 +53,7 @@ describe("react-admin-client methods used through node", () => {
     });
   });
 
-  describe("whe reading behaviors", () => {
+  describe("when reading behaviors", () => {
     it("should return behaviors collection", async () => {
       const behaviors = await readBehaviors();
       expect(behaviors.length).toEqual(2);
@@ -79,7 +95,7 @@ describe("react-admin-client methods used through node", () => {
     });
   });
 
-  describe("whe reading settings", () => {
+  describe("when reading settings", () => {
     it("should return current behavior", async () => {
       const settings = await readSettings();
       expect(settings.behavior).toEqual("base");
@@ -92,7 +108,7 @@ describe("react-admin-client methods used through node", () => {
     });
   });
 
-  describe("whe updating settings", () => {
+  describe("when updating settings", () => {
     it("should update current behavior", async () => {
       await updateSettings({
         behavior: "user2",